create NetworkGlobalsWrapper
