+++ beacon_node/lighthouse_network/src/peer_manager/mod.rs
@@ -24,7 +24,8 @@ pub use libp2p::identity::Keypair;
 
 pub mod peerdb;
 
-use crate::peer_manager::connectivity::{Connectivity, LHNetworkGlobalsConnectivity};
+use crate::peer_manager::connectivity::Connectivity;
+use crate::peer_manager::network_globals_wrapper::LHNetworkGlobalsWrapper;
 use crate::peer_manager::peerdb::client::ClientKind;
 use libp2p::multiaddr;
 pub use peerdb::peer_info::{
@@ -42,6 +43,7 @@ use types::data_column_custody_group::{
 pub mod config;
 mod connectivity;
 mod network_behaviour;
+mod network_globals_wrapper;
 
 /// The heartbeat performs regular updates such as updating reputations and performing discovery
 /// requests. This defines the interval in seconds.
@@ -117,7 +119,7 @@ pub struct PeerManager<E: EthSpec> {
     /// The logger associated with the `PeerManager`.
     log: slog::Logger,
 
-    connectivity: Connectivity<LHNetworkGlobalsConnectivity<E>>,
+    connectivity: Connectivity<LHNetworkGlobalsWrapper<E>>,
 }
 
 /// The events that the `PeerManager` outputs (requests).
@@ -204,7 +206,7 @@ impl<E: EthSpec> PeerManager<E> {
                 MIN_OUTBOUND_ONLY_FACTOR,
                 TARGET_OUTBOUND_ONLY_FACTOR,
                 discovery_enabled,
-                LHNetworkGlobalsConnectivity::new(network_globals),
+                LHNetworkGlobalsWrapper::new(network_globals),
             ),
         })
     }
